requirements.txt
tensorflow==2.6.0
numpy==1.19.5
python_telegram_bot==13.7
Pillow==8.3.2
python-dotenv==0.19.0
telegram==0.0.1
